+++ eslint-config-module/index.js
@@ -1,3 +1,9 @@
+/**
+ * Eslint Config Module
+ * -----------------------------------------------------------------------------
+ * We Make Websites eslint configuration for non-webpack projects.
+ *
+ */
 const merge = require('merge')
 
 module.exports = {
@@ -8,6 +14,7 @@ module.exports = {
   extends: [
     'eslint:recommended',
     'plugin:vue/vue3-recommended',
+    'plugin:import/recommended',
   ],
   globals: {
     geoip2: true,
@@ -16,10 +23,14 @@ module.exports = {
     Shopify: true,
     URLSearchParams: true,
   },
+  parserOptions: {
+    ecmaVersion: 2020,
+  },
   plugins: [],
   root: true,
   rules: merge(
     require('./rules/formatting'),
+    require('./rules/import'),
     require('./rules/problems'),
     require('./rules/suggestions'),
     require('./rules/vue'),