+++ canvas-component-tools/CHANGELOG.md
@@ -2,6 +2,34 @@
 
 Canvas Component Tools uses [semantic versioning](https://semver.org/).
 
+## 1.8.0 - 2023-09-19
+
+* Added tracking for logging commands and error analytics
+
+## 1.7.0 - 2023-08-23
+
+* Updated dependencies
+
+## 1.6.0 - 2023-08-22
+
+* Removed `aria-labelledby` attributes from default template `<section>` elements
+* Removed `id` attributes from default template titles
+
+## 1.5.0 - 2023-07-18
+
+* Updated Liquid templates to include _.min_ extension
+
+## 1.4.1 - 2023-07-03
+
+* Fixed global limited interactivity template causing compile errors
+
+## 1.4.0 - 2023-05-30
+
+* Updated `disabled_on` property to be disabled on all section groups
+* Fixed description not wrapping in Vue template causing eslint errors
+* Fixed description not being formatted correctly
+* Updated dependencies
+
 ## 1.3.0 - 2023-04-11
 
 * Added `disabled_on` property to async schema files to disable in header and footer section groups by default